Crate stahl_context

The context of the compiler and interpreter.

This should be your main interface to the compiler.

Structs

Context
The context in which compilation and interpretation occur.

DefContext
A single def's portion of the context.

DefTyCtorsContext
A single defty's portion of the context, while defining the constructors of the type.

DefTyKindContext
A single defty's portion of the context, while defining the kind of the type.

LibContext
A single library's portion of the context.

ModContext
A single module's portion of the context.

UnifEffs
A set of effects which can undergo unification.

Zipper
A zipper for an expression.

Enums

UnifExpr
An expression during unification. This is similar to the AST expression, but has an additional constructor for unification variables.