Struct stahl_context::Context

pub struct Context {
    pub search_paths: Vec<SharedPath>,
    // some fields omitted
}

The context in which compilation and interpretation occur.

Fields
search_paths: Vec<SharedPath>

The paths searched when looking for libraries.

Methods

impl Context

pub fn new<II>(load_std: bool, search_paths: II) -> Result<Context> where
II: IntoIterator<Item = SharedPath>,

Creates a new Context.

pub fn create_lib(
&mut self,
name: LibName,
deps: HashMap<SharedString, LibName>,
path: Option<SharedPath>
) -> LibContext

Creates a context for the library with the given name and version.

pub fn load_lib(&mut self, name: LibName) -> Result<()>

Loads the library with the given name from the search path.

pub fn load_lib_highest_version(
&mut self,
name: SharedString
) -> Result<LibName>

Loads the library with the highest version and the given name from the search path, Returns the name of the library loaded.

pub fn with_lib<'c, F>(
&'c mut self,
name: LibName,
deps: HashMap<SharedString, LibName>,
path: Option<SharedPath>,
f: F
) -> Result<&'c mut Context> where
F: FnOnce(&mut LibContext<'c>) -> Result<()>,

Runs the given closure with a LibContext, cleaning up afterwards.

pub fn insert_lib(&mut self, library: Library) -> Result<()>

Inserts the given library, creating an entry for it in the context.

pub fn get_decl(
&self,
name: FQName,
enforce_exports: bool
) -> Result<(FQName, &Decl)>

Returns the declaration referenced by the given fully qualified name. If enforce_exports is false, the decl is returned regardless of whether it is public.

pub fn get_module(
&self,
lib_name: LibName,
mod_name: SharedString
) -> Result<&Module>

Returns the module given by the given library name and module name.

pub fn std(&self) -> Option<LibName>

Returns the name of the standard library, if it has been loaded.

pub fn prelude_exports(&self) -> Option<&HashSet<SharedString>>

Returns the exports of the prelude module, if one exists.