Struct stahl_context::ModContext

pub struct ModContext<'l, 'c> { /* fields omitted */ }

A single module's portion of the context.

Methods

impl<'_, '_> ModContext<'_, '_>

pub fn elab(
&mut self,
cst_expr: &CstExpr,
ty: &CstExpr
) -> Result<(Arc<Expr>, Arc<Expr>)>

Elaborates a CST expression into an AST expression.

pub fn cst_to_unif(
&mut self,
expr: &CstExpr,
locals: &mut Vec<SharedString>
) -> Result<Rc<UnifExpr>>

Converts a CST expression to a UnifExpr, with the given variables treated as locals.

pub fn unify_ty_expr(
&self,
ty: &mut Rc<UnifExpr>,
expr: &mut Rc<UnifExpr>,
env: &mut Vec<(SharedString, Rc<UnifExpr>, Option<Rc<UnifExpr>>)>
) -> Result<()>

Unifies a (top-level) expression with its type.

impl<'l, 'c: 'l> ModContext<'l, 'c>

pub fn finish(self) -> Result<&'l mut LibContext<'c>>

Adds this module to the library.

pub fn discard(self)

Discards this module. Use this instead of just dropping in the case where the module should not be added!

pub fn add(&mut self, decl: Decl) -> Result<()>

Adds a new declaration.

pub fn add_cst_decl(&mut self, decl: CstDecl) -> Result<()>

Adds a new CST declaration.

pub fn create_def<'m>(
&'m mut self,
loc: Location,
name: SharedString
) -> DefContext<'m, 'l, 'c>

Begins creating a def.

pub fn create_defty<'m>(
&'m mut self,
loc: Location,
name: SharedString
) -> DefTyKindContext<'m, 'l, 'c>

Begins creating a defty.

pub fn with_def<'m, F>(
&'m mut self,
loc: Location,
name: SharedString,
f: F
) -> Result<&'m mut ModContext<'l, 'c>> where
F: FnOnce(&mut DefContext<'m, 'l, 'c>) -> Result<()>,

Runs the given closure with a DefContext, cleaning up afterwards.

pub fn with_defty<'m, F1, F2>(
&'m mut self,
loc: Location,
name: SharedString,
f1: F1,
f2: F2
) -> Result<&'m mut ModContext<'l, 'c>> where
F1: FnOnce(&mut DefTyKindContext<'m, 'l, 'c>) -> Result<()>,
F2: FnOnce(&mut DefTyCtorsContext<'m, 'l, 'c>) -> Result<()>,

Runs the given closure with a DefTyKindContext, then another with the corresponding DefTyCtorsContext, cleaning up afterwards.

pub fn get_decl(&self, name: FQName) -> Option<(FQName, &Decl)>

Returns the decl referred to by the given global name.

pub fn local_name_iter<'m>(
&'m self
) -> impl 'm + Iterator<Item = SharedString>

Returns an iterator over names that are locally visible.

pub fn resolve(&self, name: SharedString) -> Result<(FQName, &Decl)>

Resolves the name of a definition in the module context, returning its declaration.

pub fn add_prelude_import(&mut self, strict: bool) -> Result<()>

Adds an import for the prelude, if it exists. If strict is true, an error will be returned if the library does not import std.