Struct stahl_context::Zipper

pub struct Zipper { /* fields omitted */ }

A zipper for an expression.

Methods

impl Zipper

pub fn go_left(&mut self) -> bool

Goes "left," returning whether this was successful. If false is returned, the position will be unchanged.

pub fn go_right(&mut self) -> bool

Goes "right," returning whether this was successful. If false is returned, the position will be unchanged.

pub fn go_down(&mut self) -> bool

Goes down, preferring the leftmost possiblity, returning false if it is not possible.

pub fn go_to_leftmost_hole(&mut self) -> bool

Goes to the leftmost hole, returning whether it was possible. If not, returns false, and is at the rightmost leaf.

pub fn go_to_top(&mut self)

Goes to the top.

impl Zipper

pub fn intros(&mut self, loc: Location, vars: Vec<SharedString>)

Introduces variables into the context via a lambda expression. This may shadow other variables.

pub fn intros_pi(&mut self, loc: Location, vars: Vec<SharedString>)

Introduces variables into the context via a pi expression. This may shadow other variables. Note that there will additional holes created to the left of the currently focused expression, for the type of each new argument.

pub fn fill(&mut self, expr: Rc<UnifExpr>)

Replaces the currently focused expression, warning if it is not a hole.

impl Zipper

pub fn expr(&self) -> &UnifExpr

Returns the expression under focus.

pub fn at_hole(&self) -> bool

Returns whether the zipper is at a hole.

pub fn at_leaf(&self) -> bool

Returns whether the zipper is at a leaf.

pub fn at_top(&self) -> bool

Returns whether the zipper is at the top of the expression tree.

pub fn go_up(&mut self)

Goes up one level, panicking if we're already at the top.

pub fn go_to_call_arg(&mut self, argn: usize)

Descends to the argument of the currently selected call expression. Panics if there aren't enough arguments, or if the expression isn't a call.

pub fn go_to_call_func(&mut self)

Descends to the function of the currently selected call expression. Panics if the expression isn't a call.

pub fn go_to_lam_expr(&mut self, exprn: usize)

Descends to the expression of the currently selected lambda expression's body. Panics if the expression isn't a lambda.

pub fn go_to_lam_ty(&mut self, exprn: usize)

Descends to the type of the currently selected lambda expression's body. Panics if the expression isn't a lambda.

pub fn go_to_pi_arg(&mut self, argn: usize)

Descends to the argument of the currently selected pi expression. Panics if the expression isn't a pi.

pub fn go_to_pi_return(&mut self)

Descends to the return of the currently selected pi expression. Panics if the expression isn't a pi.